requirements.txt
# Deep Learning Framework
torch>=2.0.0
torchvision>=0.15.0

# Data Processing & Scientific Computing
numpy>=1.21.0
scipy>=1.7.0
pandas>=1.3.0
pillow>=8.3.0

# Visualization & Logging
matplotlib>=3.4.0
tqdm>=4.62.0
tensorboard>=2.7.0

# Image Processing
opencv-python>=4.5.0
scikit-image>=0.18.0

# Metrics
lpips>=0.1.4  # Learning-based image quality metrics
pytorch-fid>=0.2.0  # FID score calculation

# Utilities
pyyaml>=5.4.0  # Para configuraciones
wandb>=0.12.0  # (opcional) Para tracking de experimentos
einops>=0.4.0  # Para manipulaciones de tensores
joblib>=1.1.0  # Para paralelización

# Development Tools
pytest>=6.2.0  # Para testing
black>=21.6b0  # Para formateo de código
isort>=5.9.0   # Para ordenar imports
flake8>=3.9.0  # Para linting

# CUDA toolkit (comentado porque se instala con PyTorch)
# cudatoolkit>=11.3

# Opcional pero recomendado para mejor rendimiento
ninja>=1.10.0  # Para compilación más rápida